+++ Nursery/C/Expressions/NUCExpressionResult.m
@@ -3,21 +3,20 @@
 //  Nursery
 //
 //  Created by TAKATA Akifumi on 2024/01/06.
-//  Copyright © 2024 com.lily-bud. All rights reserved.
 //
 
 #import "NUCExpressionResult.h"
 
 @implementation NUCExpressionResult
 
-- (instancetype)initWithIntValue:(NSInteger)aValue
+- (instancetype)initWithIntValue:(int)aValue
 {
     intValue = aValue;
     
     return self;
 }
 
-- (NSInteger)intValue
+- (int)intValue
 {
     return intValue;
 }
